# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Order Management API",
        "description": "E-Commerce Order Management System API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/products/import": {
            "post": {
                "tags": [
                    "Product Import"
                ],
                "summary": "Import products from CSV file",
                "operationId": "7518bee25f8017ba79c7d3540698b55a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "csv_file"
                                ],
                                "properties": {
                                    "csv_file": {
                                        "description": "CSV file containing product data",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Products imported successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Import completed"
                                        },
                                        "results": {
                                            "properties": {
                                                "success": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "errors": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Row error: The sku field must be unique."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized - Vendor or Admin role required"
                    },
                    "422": {
                        "description": "Validation error or import failed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/search/products": {
            "get": {
                "tags": [
                    "Search"
                ],
                "summary": "Search products",
                "operationId": "7610860f1e4fb6b72f7f5a97d5815444",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Search query (minimum 2 characters)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "minLength": 2
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Products found successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error - search query too short"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Product": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "iPhone 15 Pro"
                    },
                    "description": {
                        "type": "string",
                        "example": "Latest Apple smartphone"
                    },
                    "sku": {
                        "type": "string",
                        "example": "IPHONE-15-PRO-256"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 1199.99
                    },
                    "stock_quantity": {
                        "type": "integer",
                        "example": 100
                    },
                    "low_stock_threshold": {
                        "type": "integer",
                        "example": 10
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}